Move from 'av stack branch' to 'av branch'
