+++ demos/stack_switch/stack_switch_non_managed_branch.tape
@@ -11,18 +11,18 @@ Hide
 Type '. ../test_utils.sh' Enter
 Type 'new_temp_repo' Enter
 
-Type 'av stack branch stack-1' Enter
+Type 'av branch stack-1' Enter
 Type 'create_commit "my-file" "1a" "Commit 1a"' Enter
 
-Type 'av stack branch stack-2' Enter
+Type 'av branch stack-2' Enter
 Type 'create_commit "my-file" "2a" "Commit 2a"' Enter
 
-Type 'av stack branch stack-3' Enter
+Type 'av branch stack-3' Enter
 Type 'create_commit "my-file" "3a" "Commit 3a"' Enter
 
 Type 'git switch stack-1' Enter
 
-Type 'av stack branch stack-4' Enter
+Type 'av branch stack-4' Enter
 Type 'create_commit "my-file" "4a" "Commit 4a"' Enter
 
 Type 'git switch stack-1' Enter
