+++ e2e_tests/stack_adopt_test.go
@@ -40,7 +40,7 @@ func TestStackAdopt_Success_WithStackRoot(t *testing.T) {
 
 	// main: root_commit
 	// stack-1: root_commit -> 1a -> 1b
-	RequireAv(t, "stack", "branch", "stack-1")
+	RequireAv(t, "branch", "stack-1")
 	repo.CommitFile(t, "my-file", "1a\n", gittest.WithMessage("Commit 1a"))
 	repo.CommitFile(t, "my-file", "1b\n", gittest.WithMessage("Commit 1b"))
