+++ e2e_tests/stack_sync_all_test.go
@@ -13,11 +13,11 @@ func TestStackSyncAll(t *testing.T) {
 	Chdir(t, repo.RepoDir)
 
 	repo.Git(t, "switch", "main")
-	RequireAv(t, "stack", "branch", "stack-1")
+	RequireAv(t, "branch", "stack-1")
 	repo.CommitFile(t, "my-file", "1a\n", gittest.WithMessage("Commit 1a"))
 
 	repo.Git(t, "switch", "main")
-	RequireAv(t, "stack", "branch", "stack-2")
+	RequireAv(t, "branch", "stack-2")
 	repo.CommitFile(t, "my-file", "2a\n", gittest.WithMessage("Commit 2a"))
 
 	repo.Git(t, "switch", "main")
