+++ e2e_tests/stack_tree_test.go
@@ -10,14 +10,14 @@ func TestStackTree(t *testing.T) {
 	repo := gittest.NewTempRepo(t)
 	Chdir(t, repo.RepoDir)
 
-	RequireAv(t, "stack", "branch", "foo")
+	RequireAv(t, "branch", "foo")
 	repo.CommitFile(t, "foo", "foo")
 
-	RequireAv(t, "stack", "branch", "bar")
+	RequireAv(t, "branch", "bar")
 	repo.CommitFile(t, "bar", "bar")
 
 	repo.CheckoutBranch(t, "refs/heads/main")
-	RequireAv(t, "stack", "branch", "spam")
+	RequireAv(t, "branch", "spam")
 	repo.CommitFile(t, "spam", "spam")
 
 	RequireAv(t, "stack", "tree")
